refactor components folders
