+++ components/cards/SessionCard.tsx
@@ -12,24 +12,20 @@ import {
 } from '@mui/material';
 import Box from '@mui/material/Box';
 import { useTranslations } from 'next-intl';
-import Image from 'next/image';
 import { useState } from 'react';
 import { StoryData } from 'storyblok-js-client';
-import { PROGRESS_STATUS } from '../constants/enums';
-import { rowStyle } from '../styles/common';
+import { PROGRESS_STATUS } from '../../constants/enums';
+import { rowStyle } from '../../styles/common';
 
 const cardStyle = {
   alignSelf: 'flex-start',
-  width: '100%',
+  width: { xs: '100%', md: 'calc(50% - 1rem)' },
   backgroundColor: 'background.default',
 } as const;
 
 const cardContentStyle = {
-  ...rowStyle,
-  gap: 2,
   padding: { xs: 2, md: 3 },
   paddingBottom: { xs: 1, md: 1 },
-  minHeight: { xs: 124, md: 136 },
 } as const;
 
 const collapseContentStyle = {
@@ -43,74 +39,70 @@ const cardActionsStyle = {
   alignItems: 'end',
 } as const;
 
-const rowStyles = {
+const cardContentRowStyles = {
   ...rowStyle,
   gap: 1.5,
 } as const;
 
-interface CourseCardProps {
-  course: StoryData;
-  courseProgress: PROGRESS_STATUS | null;
+interface SessionCardProps {
+  session: StoryData;
+  sessionProgress: PROGRESS_STATUS;
 }
 
-const CourseCard = (props: CourseCardProps) => {
-  const { course, courseProgress } = props;
+const SessionCard = (props: SessionCardProps) => {
+  const { session, sessionProgress } = props;
   const [expanded, setExpanded] = useState<boolean>(false);
+
   const t = useTranslations('Courses');
 
   const handleExpandClick = () => {
     setExpanded(!expanded);
   };
 
+  const arrowStyle = {
+    transform: expanded ? 'rotate(180deg)' : 'none',
+  } as const;
+
   return (
-    <Card sx={cardStyle} key={course.id}>
+    <Card sx={cardStyle} key={session.id}>
       <CardActionArea
-        href={`/${course.full_slug}`}
-        aria-label={`${t('navigateToCourse')} ${course.content.name}`}
+        href={`/${session.full_slug}`}
+        aria-label={`${t('navigateToSession')} ${session.name}`}
       >
         <CardContent sx={cardContentStyle}>
-          <Image
-            alt={course.content.image.alt}
-            src={course.content.image.filename}
-            width={100}
-            height={100}
-            objectFit="contain"
-          />
-          <Box flex={1}>
-            <Typography key={course.slug} component="h3" variant="h3">
-              {course.content.name}
-            </Typography>
-            {courseProgress !== PROGRESS_STATUS.NOT_STARTED && (
-              <Box sx={rowStyles}>
-                {courseProgress === PROGRESS_STATUS.STARTED && (
-                  <DonutLargeIcon color="error"></DonutLargeIcon>
-                )}
-                {courseProgress === PROGRESS_STATUS.COMPLETED && (
-                  <CheckCircleIcon color="error"></CheckCircleIcon>
-                )}
-                <Typography>{courseProgress}</Typography>
+          <Box sx={cardContentRowStyles}>
+            {sessionProgress !== PROGRESS_STATUS.NOT_STARTED && (
+              <Box mt={0.5}>
+                {sessionProgress === PROGRESS_STATUS.STARTED && <DonutLargeIcon color="error" />}
+                {sessionProgress === PROGRESS_STATUS.COMPLETED && <CheckCircleIcon color="error" />}
               </Box>
             )}
+            <Typography component="h3" variant="h3">
+              {session.content.name}
+            </Typography>
           </Box>
+          <Typography color="grey.700">
+            {t('session')} {session.position / 10 - 1}
+          </Typography>
         </CardContent>
       </CardActionArea>
       <CardActions sx={cardActionsStyle}>
         <IconButton
           sx={{ marginLeft: 'auto' }}
-          aria-label={`${t('expandSummary')} ${course.content.name}`}
+          aria-label={`${t('expandSummary')} ${session.name}`}
           onClick={handleExpandClick}
           size="small"
         >
-          <KeyboardArrowDownIcon color="error"></KeyboardArrowDownIcon>
+          <KeyboardArrowDownIcon sx={arrowStyle} color="error" />
         </IconButton>
       </CardActions>
       <Collapse in={expanded} timeout="auto" unmountOnExit>
         <CardContent sx={collapseContentStyle}>
-          <Typography paragraph>{course.content.description}</Typography>
+          <Typography>{session.content.description}</Typography>
         </CardContent>
       </Collapse>
     </Card>
   );
 };
 
-export default CourseCard;
+export default SessionCard;
