+++ components/forms/RegisterForm.tsx
@@ -10,12 +10,11 @@ import { useTranslations } from 'next-intl';
 import { useRouter } from 'next/router';
 import * as React from 'react';
 import { useEffect, useState } from 'react';
-import { useAddUserMutation, useValidateCodeMutation } from '../app/api';
-import { setUserToken } from '../app/userSlice';
-import Link from '../components/Link';
-import { auth } from '../config/firebase';
-import rollbar from '../config/rollbar';
-import { LANGUAGES, PARTNER_ACCESS_CODE_STATUS } from '../constants/enums';
+import { useAddUserMutation, useValidateCodeMutation } from '../../app/api';
+import { setUserToken } from '../../app/userSlice';
+import { auth } from '../../config/firebase';
+import rollbar from '../../config/rollbar';
+import { LANGUAGES, PARTNER_ACCESS_CODE_STATUS } from '../../constants/enums';
 import {
   REGISTER_ERROR,
   REGISTER_FIREBASE_ERROR,
@@ -24,11 +23,12 @@ import {
   VALIDATE_ACCESS_CODE_INVALID,
   VALIDATE_ACCESS_CODE_REQUEST,
   VALIDATE_ACCESS_CODE_SUCCESS,
-} from '../constants/events';
-import { Partner } from '../constants/partners';
-import { useAppDispatch } from '../hooks/store';
-import { getErrorMessage } from '../utils/errorMessage';
-import logEvent, { getEventUserData } from '../utils/logEvent';
+} from '../../constants/events';
+import { Partner } from '../../constants/partners';
+import { useAppDispatch } from '../../hooks/store';
+import { getErrorMessage } from '../../utils/errorMessage';
+import logEvent, { getEventUserData } from '../../utils/logEvent';
+import Link from '../common/Link';
 
 const containerStyle = {
   marginY: 3,
