+++ components/layout/NavigationDrawer.tsx
@@ -7,9 +7,9 @@ import Drawer from '@mui/material/Drawer';
 import { useTranslations } from 'next-intl';
 import Image from 'next/image';
 import * as React from 'react';
+import { columnStyle } from '../../styles/common';
+import Link from '../common/Link';
 import bloomLogo from '../public/bloom_logo.svg';
-import { columnStyle } from '../styles/common';
-import Link from './Link';
 import NavigationMenu from './NavigationMenu';
 
 const drawerContainerStyle = {
