+++ components/layout/PartnerHeader.tsx
@@ -3,7 +3,7 @@ import Container from '@mui/material/Container';
 import { useTranslations } from 'next-intl';
 import Image from 'next/image';
 import * as React from 'react';
-import { rowStyle } from '../styles/common';
+import { rowStyle } from '../../styles/common';
 
 const headerContainerStyles = {
   ...rowStyle,
