+++ components/layout/TopBar.tsx
@@ -6,11 +6,11 @@ import { useTranslations } from 'next-intl';
 import Image from 'next/image';
 import { useRouter } from 'next/router';
 import * as React from 'react';
-import { RootState } from '../app/store';
-import { useTypedSelector } from '../hooks/store';
+import { RootState } from '../../app/store';
+import { useTypedSelector } from '../../hooks/store';
+import { rowStyle } from '../../styles/common';
+import Link from '../common/Link';
 import bloomLogo from '../public/bloom_logo.svg';
-import { rowStyle } from '../styles/common';
-import Link from './Link';
 import NavigationDrawer from './NavigationDrawer';
 import NavigationMenu from './NavigationMenu';
 import UserMenu from './UserMenu';
