+++ components/storyblok/Faqs.tsx
@@ -5,11 +5,11 @@ import AccordionSummary from '@mui/material/AccordionSummary';
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
 import { useTranslations } from 'next-intl';
-import { THERAPY_FAQ_OPENED } from '../constants/events';
-import { faqItem } from '../constants/faqs';
-import { Partner } from '../constants/partners';
-import logEvent from '../utils/logEvent';
-import Link from './Link';
+import { THERAPY_FAQ_OPENED } from '../../constants/events';
+import { faqItem } from '../../constants/faqs';
+import { Partner } from '../../constants/partners';
+import logEvent from '../../utils/logEvent';
+import Link from '../common/Link';
 
 interface FaqsProps {
   translations: string;
