+++ pages/404.tsx
@@ -4,8 +4,8 @@ import { GetStaticPropsContext, NextPage } from 'next';
 import Image from 'next/image';
 import { useTranslations } from 'use-intl';
 import { RootState } from '../app/store';
-import Link from '../components/Link';
-import LoadingContainer from '../components/LoadingContainer';
+import Link from '../components/common/Link';
+import LoadingContainer from '../components/common/LoadingContainer';
 import { useTypedSelector } from '../hooks/store';
 import bloomHead from '../public/illustration_bloom_head.svg';
 import { columnStyle } from '../styles/common';
