+++ pages/_app.tsx
@@ -10,9 +10,9 @@ import { useEffect } from 'react';
 import { hotjar } from 'react-hotjar';
 import { wrapper } from '../app/store';
 import { setUserLoading, setUserToken } from '../app/userSlice';
-import Footer from '../components/Footer';
-import LeaveSiteButton from '../components/LeaveSiteButton';
-import TopBar from '../components/TopBar';
+import Footer from '../components/layout/Footer';
+import LeaveSiteButton from '../components/layout/LeaveSiteButton';
+import TopBar from '../components/layout/TopBar';
 import createEmotionCache from '../config/emotionCache';
 import { auth } from '../config/firebase';
 import { useAppDispatch } from '../hooks/store';
