+++ pages/auth/login.tsx
@@ -10,9 +10,9 @@ import { GetStaticPropsContext } from 'next';
 import { useTranslations } from 'next-intl';
 import Head from 'next/head';
 import Image from 'next/image';
-import Link from '../../components/Link';
-import LoginForm from '../../components/LoginForm';
-import PartnerHeader from '../../components/PartnerHeader';
+import Link from '../../components/common/Link';
+import LoginForm from '../../components/forms/LoginForm';
+import PartnerHeader from '../../components/layout/PartnerHeader';
 import { getAllPartnersContent } from '../../constants/partners';
 import illustrationBloomHeadYellow from '../../public/illustration_bloom_head_yellow.svg';
 import illustrationLeafMix from '../../public/illustration_leaf_mix.svg';
