+++ pages/auth/register.tsx
@@ -13,9 +13,9 @@ import Image from 'next/image';
 import { useRouter } from 'next/router';
 import * as React from 'react';
 import { useEffect, useState } from 'react';
-import Link from '../../components/Link';
-import PartnerHeader from '../../components/PartnerHeader';
-import RegisterForm from '../../components/RegisterForm';
+import Link from '../../components/common/Link';
+import RegisterForm from '../../components/forms/RegisterForm';
+import PartnerHeader from '../../components/layout/PartnerHeader';
 import { getAllPartnersContent, getPartnerContent, Partner } from '../../constants/partners';
 import illustrationBloomHeadYellow from '../../public/illustration_bloom_head_yellow.svg';
 import illustrationLeafMix from '../../public/illustration_leaf_mix.svg';
