+++ pages/auth/reset-password.tsx
@@ -7,8 +7,8 @@ import type { NextPage } from 'next';
 import { GetStaticPropsContext } from 'next';
 import { useTranslations } from 'next-intl';
 import { useRouter } from 'next/router';
-import PartnerHeader from '../../components/PartnerHeader';
-import { EmailForm, PasswordForm } from '../../components/ResetPasswordForm';
+import { EmailForm, PasswordForm } from '../../components/forms/ResetPasswordForm';
+import PartnerHeader from '../../components/layout/PartnerHeader';
 import illustrationBloomHeadYellow from '../../public/illustration_bloom_head_yellow.svg';
 import welcomeToBloom from '../../public/welcome_to_bloom.svg';
 import { rowStyle } from '../../styles/common';
