+++ pages/courses/[slug].tsx
@@ -8,10 +8,10 @@ import { StoriesParams, StoryData } from 'storyblok-js-client';
 import { render } from 'storyblok-rich-text-react-renderer';
 import { Course } from '../../app/coursesSlice';
 import { RootState } from '../../app/store';
-import CourseIntroduction from '../../components/CourseIntroduction';
-import Header from '../../components/Header';
-import Link from '../../components/Link';
-import SessionCard from '../../components/SessionCard';
+import SessionCard from '../../components/cards/SessionCard';
+import Link from '../../components/common/Link';
+import Header from '../../components/layout/Header';
+import CourseIntroduction from '../../components/video/CourseIntroduction';
 import Storyblok, { useStoryblok } from '../../config/storyblok';
 import { LANGUAGES, PROGRESS_STATUS } from '../../constants/enums';
 import { COURSE_OVERVIEW_VIEWED } from '../../constants/events';
