+++ pages/courses/[slug]/[sessionSlug].tsx
@@ -16,12 +16,12 @@ import { render } from 'storyblok-rich-text-react-renderer';
 import { useCompleteSessionMutation, useStartSessionMutation } from '../../../app/api';
 import { Course, Session } from '../../../app/coursesSlice';
 import { RootState } from '../../../app/store';
-import CrispButton from '../../../components/CrispButton';
-import Header from '../../../components/Header';
-import Link from '../../../components/Link';
-import SessionContentCard from '../../../components/SessionContentCard';
-import Video from '../../../components/Video';
-import VideoTranscriptModal from '../../../components/VideoTranscriptModal';
+import SessionContentCard from '../../../components/cards/SessionContentCard';
+import Link from '../../../components/common/Link';
+import CrispButton from '../../../components/crisp/CrispButton';
+import Header from '../../../components/layout/Header';
+import Video from '../../../components/video/Video';
+import VideoTranscriptModal from '../../../components/video/VideoTranscriptModal';
 import rollbar from '../../../config/rollbar';
 import Storyblok, { useStoryblok } from '../../../config/storyblok';
 import { LANGUAGES, PROGRESS_STATUS } from '../../../constants/enums';
