+++ pages/courses/index.tsx
@@ -7,9 +7,9 @@ import Head from 'next/head';
 import { useEffect, useState } from 'react';
 import { StoryData } from 'storyblok-js-client';
 import { RootState } from '../../app/store';
-import CourseCard from '../../components/CourseCard';
-import Header from '../../components/Header';
-import LoadingContainer from '../../components/LoadingContainer';
+import CourseCard from '../../components/cards/CourseCard';
+import LoadingContainer from '../../components/common/LoadingContainer';
+import Header from '../../components/layout/Header';
 import Storyblok from '../../config/storyblok';
 import { PROGRESS_STATUS } from '../../constants/enums';
 import { COURSE_LIST_VIEWED } from '../../constants/events';
