+++ pages/partner-admin/create-access-code.tsx
@@ -6,8 +6,8 @@ import { useTranslations } from 'next-intl';
 import Head from 'next/head';
 import { useEffect } from 'react';
 import { RootState } from '../../app/store';
-import CreateAccessCodeForm from '../../components/CreateAccessCodeForm';
-import AdminHeader from '../../components/PartnerAdminHeader';
+import CreateAccessCodeForm from '../../components/forms/CreateAccessCodeForm';
+import AdminHeader from '../../components/layout/PartnerAdminHeader';
 import { CREATE_PARTNER_ACCESS_VIEWED } from '../../constants/events';
 import { useTypedSelector } from '../../hooks/store';
 import bloomLogo from '../../public/bloom_logo.svg';
