+++ pages/therapy/book-session.tsx
@@ -10,9 +10,9 @@ import Script from 'next/script';
 import { useEffect, useState } from 'react';
 import { PartnerAccess } from '../../app/partnerAccessSlice';
 import { RootState } from '../../app/store';
-import Faqs from '../../components/Faqs';
-import Header from '../../components/Header';
-import ImageTextGrid, { ImageTextItem } from '../../components/ImageTextGrid';
+import ImageTextGrid, { ImageTextItem } from '../../components/common/ImageTextGrid';
+import Header from '../../components/layout/Header';
+import Faqs from '../../components/storyblok/Faqs';
 import { getSimplybookWidgetConfig } from '../../config/simplybook';
 import { THERAPY_BOOKING_OPENED, THERAPY_BOOKING_VIEWED } from '../../constants/events';
 import { therapyFaqs } from '../../constants/faqs';
