+++ pages/therapy/confirmed-session.tsx
@@ -8,9 +8,9 @@ import Image from 'next/image';
 import { useEffect, useState } from 'react';
 import { PartnerAccess } from '../../app/partnerAccessSlice';
 import { RootState } from '../../app/store';
-import Faqs from '../../components/Faqs';
-import Header from '../../components/Header';
-import Link from '../../components/Link';
+import Link from '../../components/common/Link';
+import Header from '../../components/layout/Header';
+import Faqs from '../../components/storyblok/Faqs';
 import { THERAPY_CONFIRMATION_VIEWED } from '../../constants/events';
 import { therapyFaqs } from '../../constants/faqs';
 import { useTypedSelector } from '../../hooks/store';
