+++ pages/welcome.tsx
@@ -12,9 +12,9 @@ import Head from 'next/head';
 import Image from 'next/image';
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
-import CodeForm from '../components/CodeForm';
-import Link from '../components/Link';
-import PartnerHeader from '../components/PartnerHeader';
+import Link from '../components/common/Link';
+import CodeForm from '../components/forms/CodeForm';
+import PartnerHeader from '../components/layout/PartnerHeader';
 import bloomBumbleLogo from '../public/bloom_bumble_logo.svg';
 import illustrationBloomHeadYellow from '../public/illustration_bloom_head_yellow.svg';
 import illustrationLeafMix from '../public/illustration_leaf_mix.svg';
