+++ utils/partnerAdminGuard.tsx
@@ -4,7 +4,7 @@ import Head from 'next/head';
 import Image from 'next/image';
 import { useTranslations } from 'use-intl';
 import { RootState } from '../app/store';
-import Link from '../components/Link';
+import Link from '../components/common/Link';
 import { useTypedSelector } from '../hooks/store';
 import illustrationPerson4Peach from '../public/illustration_person4_peach.svg';
 import { columnStyle } from '../styles/common';
